Detect the region number before creating the new table
In the case that the sequence needs to be pulled from an existing table in
the database, we need to detect the region number before creating a new
table (and its new sequence).  If we do it after creating the table and its
sequence, we have a chance of selecting that sequence and using it to
determine the region number, which would incorrectly be 0.